Trying things with the enemies

+++ Mega_Man_Requiem/Motor2D/j1Pathfinding.h
@@ -42,7 +42,7 @@ public:
 	bool IsWalkable(const iPoint& pos) const;
 
 	// Utility: return the walkability value of a tile
-	uchar GetTileAt(const iPoint& pos) const;
+	uint GetTileAt(const iPoint& pos) const;
 
 private:
 
@@ -50,7 +50,7 @@ private:
 	uint width;
 	uint height;
 	// all map walkability values [0..255]
-	uchar* map;
+	uint* map;
 	// we store the created path here
 	p2DynArray<iPoint> last_path;
 };
